Reorganise components, reducers and store
Consolidating all in one location
