--- web/reactjs/src/actions/Users/containers/Users.js
@@ -1,34 +1,0 @@
-// Copyright 2020 Paul Sitoh
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import { connect } from 'react-redux';
-import { Users } from '../../../presentation';
-import axios from 'axios';
-
-const mapStateToProps = (state) => {
-    return {
-        users: state.usersReducer.users
-    };
-};
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-        fetchUser: () => dispatch({type: 'FETCH_USERS', payload: axios.get('/api/users')})
-    };
-};
-
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(Users);
