--- web/reactjs/src/actions/store.js
@@ -1,26 +1,0 @@
-// Copyright 2020 Paul Sitoh
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import 'regenerator-runtime/runtime';
-import { applyMiddleware, createStore, combineReducers } from 'redux';
-import { createLogger } from 'redux-logger';
-import { createPromise } from 'redux-promise-middleware';
-import { usersReducer } from './Users/reducers';
-import { authReducer } from './Auth/reducers';
-
-export const store = createStore(
-    combineReducers({usersReducer, authReducer}),
-    {},
-    applyMiddleware(createPromise({types:{fulfilled: 'success'}}), createLogger())
-);
