+++ web/reactjs/src/modules/Auth/containers/index.js
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { default as DashboardItem } from './DashboardItem';
+export { default as Auth } from './Auth';
