+++ web/reactjs/src/modules/Auth/index.js
@@ -12,25 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React from 'react';
-
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(theme => ({
-    root: {
-        padding: theme.spacing(4)
-    }
-}));
-
-const Settings = () => {
-    const classes = useStyles();
-
-    return (
-        <div className={classes.root}>
-            
-        </div>
-    );
-
-};
-
-export default Settings;
+export { default } from './containers/Auth';
