+++ web/reactjs/src/modules/Auth/reducers/index.js
