+++ web/reactjs/src/modules/index.js
@@ -12,4 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export { default } from './Settings';
+export { default as Auth } from './Auth';
+export { default as Dashboard } from './Dashboard';
+export { default as NotFound } from './NotFound';
